+++ src/rpc_client.cpp
@@ -3,7 +3,7 @@
 #include <curl/curl.h>
 #include <tinyformat.h>
 
-#include <fstream>
+#include "http_client.h"
 
 UniValue MakeArg(std::string const& str)
 {
@@ -58,7 +58,7 @@ RPCClient::Result RPCClient::SendMethod(bool no_proxy, std::string const& method
     if (received_data.empty()) {
         throw NetError("empty result from RPC server");
     }
-    int n = received_data.size();
+    int n = static_cast<int>(received_data.size());
     received_data.resize(n + 1);
     received_data[n] = '\0';
     char const* psz = reinterpret_cast<char const*>(received_data.data());
